% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abundances}
\alias{abundances}
\title{Abundances - frequencies of frequencies}
\usage{
abundances(freqs, freqs0 = NULL)
}
\arguments{
\item{freqs}{A frequency table of data vector x.}

\item{freqs0}{A second optional frequency table to be merged to the freqs parameter.}
}
\description{
A function to calculate frequencies of frequencies or the combined
frequencies of frequencies of two vectors.
Freqs are calculated by table(x), where x is data. Freqs0 is optional. It is
used in the case of calculating abundances of test data when the frequencies of
the training data are already known. Freqs0 is table(x0) ,where x0 is training
data. Abundances of any kind of data vector x can be calculated with
table(table(x)). This returns a named vector that is used to calculate
probabilities and make classifications.
}
\examples{
set.seed(111)
x<-rpois(10,10)
abundances(table(x))

y<-rpois(2,10)
abundances(table(x), table(y))
}
\keyword{abundances}
