% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{as.spec.data.frame}
\alias{as.spec.data.frame}
\title{Coerce to Spec from Data Frame}
\usage{
\method{as.spec}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed arguments}
}
\value{
spec
}
\description{
Coerces to spec from data.frame already having basic properties.
}
\examples{
data(drug)
as.spec(specification(drug, tol = 3))
}
