% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{as.spec}
\alias{as.spec}
\title{Coerce to Spec}
\usage{
as.spec(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Coerces to class \code{spec}, a specification object
}
\seealso{
Other as.spec: 
\code{\link{as.spec.character}()},
\code{\link{as.spec.data.frame}()},
\code{\link{read.spec}()},
\code{\link{write.spec}()}
}
\concept{as.spec}
