% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{as.vector.spec}
\alias{as.vector.spec}
\title{Coerce to Vector from Spec.}
\usage{
\method{as.vector}{spec}(x, mode = "any")
}
\arguments{
\item{x}{spec}

\item{mode}{to match generic}
}
\description{
Coerces to vector from spec by returning \code{column}.
}
\concept{as.vector}
\keyword{internal}
