% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAffinity.R
\name{calculateAffinityMatrix}
\alias{calculateAffinityMatrix}
\title{Calculate the affinity matrix based on the similarity matrix}
\usage{
calculateAffinityMatrix(similarityMatrix, neighboorsNumber = 2)
}
\arguments{
\item{similarityMatrix}{Matrix of similarity between all points in the time series}

\item{neighboorsNumber}{Number of neighbors to consider affinity between nodes}
}
\value{
Affinity matrix based on the similarity matrix
}
\description{
Calculate the affinity matrix based on the similarity matrix
}
\details{
Calculate the affinity matrix based on the similarity matrix
         If the number of neighbors is equal to or greater than the 
         similarity matrix then the similarity and affinity matrix are equal
}
\author{
Luis Gustavo Uzai
}
