% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterEstimation.R
\name{clusterEstimatetNumber}
\alias{clusterEstimatetNumber}
\title{Estimate the number of possible clusters}
\usage{
clusterEstimatetNumber(eigenvectorValues, tolerance, maxClusterNumber)
}
\arguments{
\item{eigenvectorValues}{Eigenvector matrix based on the affinity matrix}

\item{tolerance}{approximation to consider valid clusters}

\item{maxClusterNumber}{maximum number of calculable clusters}
}
\value{
An estimated number of clusters
}
\description{
Adaptation of the bartlett method of the speccalt package to 
             estimate the number of clusters in the context of spectral 
             clustering to detect change points
}
\details{
Adaptation of the bartlett method of the speccalt package to 
         estimate the number of clusters in the context of spectral 
         clustering to detect change points
}
\author{
Luis Gustavo Uzai
}
