% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{codes.spec}
\alias{codes.spec}
\title{Extract Codes from Specification}
\usage{
\method{codes}{spec}(x, column = x$column, ...)
}
\arguments{
\item{x}{spec}

\item{column}{character: items in \code{column} to check}

\item{...}{passed arguments}
}
\value{
list
}
\description{
Extracts codes from specification guide column.
}
\concept{codes}
\keyword{internal}
