% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{convert_to_data_tables}
\alias{convert_to_data_tables}
\title{Converts a spectra object into a list of data.frames}
\usage{
convert_to_data_tables(this, policy_duplicate_bands = "ADD_SMALL_N", ...)
}
\arguments{
\item{this}{TODO}

\item{...}{TODO}
}
\description{
Note that there is no analogous function to this conversion. That is, you
cannot convert a list of tables or a data frame back to a spectra object.
}
\details{
* 1) The target data.table, with one row per target measurement.
* 2) The reference data.table, with the __unique__ reference measurements.
In this case, the data returned is either "radiance", "irradiance" or "counts"
}

