% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSimilarity.R
\name{convertToMatrixTimeSeries}
\alias{convertToMatrixTimeSeries}
\title{Converts the time series to position and value matrix}
\usage{
convertToMatrixTimeSeries(data)
}
\arguments{
\item{data}{List of values corresponding to the time series}
}
\value{
The key matrix and value of the time series.
}
\description{
Converts the time series to position and value matrix
}
\details{
Gets a list of values of any size and creates a key and 
         value array of all positions
}
\author{
Luis Gustavo Uzai
}
