% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylSet.R, R/rPD.R
\name{cylSet}
\alias{cylSet}
\title{Simulated approximation of rPD}
\usage{
cylSet(n, psi)

cylSet(n, psi)
}
\arguments{
\item{n}{Number of allowed unique species.}

\item{psi}{Dispersal parameter.}
}
\description{
Algorithm A from Zarepour, M., and Al Labadi, L. (2012). On a Rapid Simulation of the Dirichlet
Process. Statistics and Probability Letters, 82, 916-924.
https://www.sciencedirect.com/science/article/pii/S0167715212000302?via\%3Dihub

Algorithm A from Zarepour, M., and Al Labadi, L. (2012). On a Rapid Simulation of the Dirichlet
Process. Statistics and Probability Letters, 82, 916-924.
https://www.sciencedirect.com/science/article/pii/S0167715212000302?via\%3Dihub
}
\examples{
m<- 10000
psi<- 100
p<- cylSet(m, psi)
# Sums up to 1.
sum(p)

# drawing a sample from PD(psi) of size 1000
s<- sample.int(m, size=1000, replace = TRUE, prob = p)

# MLE based on the sample should be close to 100, the used psi.
MLEp(table(table(s)))

}
