% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{decodes.spec}
\alias{decodes.spec}
\title{Extract Decodes from Specification}
\usage{
\method{decodes}{spec}(x, column = x$column, ...)
}
\arguments{
\item{x}{spec}

\item{column}{character: items in \code{column} to check}

\item{...}{passed arguments}
}
\value{
list
}
\description{
Extracts decodes from specification guide column.
}
\concept{decodes}
\keyword{internal}
