% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spec.R
\docType{data}
\name{DEVICE1}
\alias{DEVICE1}
\title{DEVICE1}
\format{The format is:
        Value        Class
        1.063400     1
       -0.953410     1
       ...
       -0.596090     2 
       ...}
\usage{
DEVICE1
}
\description{
Derivation of RefrigerationDevices of the UCR Time Series Classification Repository
             These problems were taken from data recorded as part of government sponsored study
             called Powering the Nation.  The intention was to collect behavioural data about 
             how consumers use electricity within the home to help reduce the UK's carbon 
             footprint.
}
\keyword{datasets}
