% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_distribution.R
\name{dlPD}
\alias{dlPD}
\title{Poisson Dirichlet distribution}
\usage{
dlPD(abund, psi)
}
\arguments{
\item{abund}{An abundance vector.}

\item{psi}{Dispersal parameter. Accepted values are positive numbers, "a" for absolute value psi=1 by default, or "r" for relative value psi equals oriignal sample size.}
}
\value{
dlPD returns the log-probability of the abundance vector of the data vector x,
}
\description{
LogProbability of a data vector x from Ewens' sampling formula is given below.
accepts either a raw data vector x or its frequency vector, table(x). The higher the
dispersal parameter the psi is, the higher the amount of distinct observed
species will be. In terms of the paintbox process, a high psi increases the
size of the continuous part p_0 of the process, while a low psi will increase
the size of the discrete parts p_1, ... p_k.
}
\examples{
set.seed(111)
s <- rPD(100,5)
a=abundances(table(s))
dlPD(a, 5)

}
\keyword{Poisson-Dirichlet}
\keyword{distribution}
