% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{.guide.default}
\alias{.guide.default}
\title{Make Default Guide}
\usage{
\method{.guide}{default}(x, tol = 10, ...)
}
\arguments{
\item{x}{object}

\item{tol}{integer}

\item{...}{dots}
}
\description{
Makes default guide.
}
\seealso{
Other .guide: 
\code{\link{.guide.factor}()},
\code{\link{.guide.integer}()},
\code{\link{.guide.numeric}()},
\code{\link{.guide}()}
}
\concept{.guide}
\keyword{internal}
