% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{.guide.factor}
\alias{.guide.factor}
\title{Make Factor Guide}
\usage{
\method{.guide}{factor}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Makes Factor guide.
}
\seealso{
Other .guide: 
\code{\link{.guide.default}()},
\code{\link{.guide.integer}()},
\code{\link{.guide.numeric}()},
\code{\link{.guide}()}
}
\concept{.guide}
\keyword{internal}
