% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{.guide.integer}
\alias{.guide.integer}
\title{Make Integer Guide}
\usage{
\method{.guide}{integer}(x, tol = 10, ...)
}
\arguments{
\item{x}{object}

\item{tol}{integer}

\item{...}{dots}
}
\description{
Makes integer guide.
}
\seealso{
Other .guide: 
\code{\link{.guide.default}()},
\code{\link{.guide.factor}()},
\code{\link{.guide.numeric}()},
\code{\link{.guide}()}
}
\concept{.guide}
\keyword{internal}
