% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{.guide.numeric}
\alias{.guide.numeric}
\title{Make Numeric Guide}
\usage{
\method{.guide}{numeric}(x, digits = 20, ...)
}
\arguments{
\item{x}{object}

\item{digits}{integer}

\item{...}{dots}
}
\description{
Makes numeric guide.
}
\seealso{
Other .guide: 
\code{\link{.guide.default}()},
\code{\link{.guide.factor}()},
\code{\link{.guide.integer}()},
\code{\link{.guide}()}
}
\concept{.guide}
\keyword{internal}
