% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{.guide}
\alias{.guide}
\title{Make Guide}
\usage{
.guide(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Makes Guide
}
\seealso{
Other .guide: 
\code{\link{.guide.default}()},
\code{\link{.guide.factor}()},
\code{\link{.guide.integer}()},
\code{\link{.guide.numeric}()}
}
\concept{.guide}
\keyword{internal}
