% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{.label.default}
\alias{.label.default}
\title{Make Default Label}
\usage{
\method{.label}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Makes default label.
}
\seealso{
Other .label: 
\code{\link{.label}()}
}
\concept{.label}
\keyword{internal}
