% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{.label}
\alias{.label}
\title{Make Label}
\usage{
.label(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Makes label.
}
\seealso{
Other .label: 
\code{\link{.label.default}()}
}
\concept{.label}
\keyword{internal}
