% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{.required.default}
\alias{.required.default}
\title{Make Default Required}
\usage{
\method{.required}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Makes default required.
}
\seealso{
Other .required: 
\code{\link{.required}()}
}
\concept{.required}
\keyword{internal}
