% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{.required}
\alias{.required}
\title{Make Required}
\usage{
.required(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Makes required.
}
\seealso{
Other .required: 
\code{\link{.required.default}()}
}
\concept{.required}
\keyword{internal}
