% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{.type.default}
\alias{.type.default}
\title{Make Default Type}
\usage{
\method{.type}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Makes default type.
}
\seealso{
Other .type: 
\code{\link{.type.numeric}()},
\code{\link{.type.timepoint}()},
\code{\link{.type}()}
}
\concept{.type}
\keyword{internal}
