% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{.type.numeric}
\alias{.type.numeric}
\title{Make Numeric Type}
\usage{
\method{.type}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Makes numeric type.
}
\seealso{
Other .type: 
\code{\link{.type.default}()},
\code{\link{.type.timepoint}()},
\code{\link{.type}()}
}
\concept{.type}
\keyword{internal}
