% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{.type}
\alias{.type}
\title{Make Type}
\usage{
.type(x, ...)
}
\description{
Makes type
}
\seealso{
Other .type: 
\code{\link{.type.default}()},
\code{\link{.type.numeric}()},
\code{\link{.type.timepoint}()}
}
\concept{.type}
\keyword{internal}
