% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{.type.timepoint}
\alias{.type.timepoint}
\title{Make Timepoint Type}
\usage{
\method{.type}{timepoint}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\description{
Makes timepoint type.
}
\seealso{
Other .type: 
\code{\link{.type.default}()},
\code{\link{.type.numeric}()},
\code{\link{.type}()}
}
\concept{.type}
\keyword{internal}
