% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPD.pred.R
\name{dPD.pred}
\alias{dPD.pred}
\title{Updated predictive density from Ewens sampling formula}
\usage{
dPD.pred(x, psi, freqs0, coeffs, tr.prob)
}
\description{
Calculates predictive probability conditional on training data probability.
x is test data, other parameters come  from training data object.
coeffs avoids calculating n!/Stirlings(psi, n) again for training data.
}
