% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{encoded.spec}
\alias{encoded.spec}
\title{Checks whether a Specification is Encoded}
\usage{
\method{encoded}{spec}(x, column = x$column, ...)
}
\arguments{
\item{x}{spec}

\item{column}{character: items in \code{column} to check}

\item{...}{passed arguments}
}
\value{
logical with same length as \code{column} argument
}
\description{
Checks whether a specification is encoded with respect to the guide column.
}
\keyword{internal}
