% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spec.R
\docType{data}
\name{FTIR1}
\alias{FTIR1}
\title{FTIR1}
\format{The format is:
        Value        Class
        1.063400     1
       -0.953410     1
       ...
       -0.596090     2 
       ...}
\usage{
FTIR1
}
\description{
Derivation of Meat of the UCR Time Series Classification Repository
             Food spectrographs are used in chemometrics to classify food types, a task that has 
             obvious applications in food safety and quality assurance. The classes are chicken, 
             pork and turkey.
}
\keyword{datasets}
