% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSimilarity.R
\name{gaussianKernel}
\alias{gaussianKernel}
\title{Calculate Gaussian Kernel}
\usage{
gaussianKernel(x1, x2, alpha = 1)
}
\arguments{
\item{x1}{first valor to computate}

\item{x2}{second valor to computate}

\item{alpha}{Alpha Measure}
}
\value{
Measure of similarity between two points.
}
\description{
Measure of similarity between two points represented by x1 and x2
}
\details{
Measure of similarity between two points represented by x1 and x2
}
\author{
Luis Gustavo Uzai
}
