% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateEigenvector.R
\name{generateEigenvectorMatrix}
\alias{generateEigenvectorMatrix}
\title{Calculate the eigenvector of the affinity matrix}
\usage{
generateEigenvectorMatrix(affinityMatrix)
}
\arguments{
\item{affinityMatrix}{Affinity matrix based on the similarity matrix 
based on key and value matrix of the  time series}
}
\value{
Eigenvector matrix based on the affinity matrix
}
\description{
Calculate the eigenvector of the affinity matrix
}
\details{
Calculates the laplacian matrix based on the affinity matrix and 
         calculates the auto values of the graph with the eigen function
}
\author{
Luis Gustavo Uzai
}
