% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSimilarity.R
\name{generateSimilarityMatrix}
\alias{generateSimilarityMatrix}
\title{Calculate Similarity Matrix}
\usage{
generateSimilarityMatrix(data, similarityMeasure)
}
\arguments{
\item{data}{Key and value matrix of a time series}

\item{similarityMeasure}{Measure of similarity between two points represented by x1 and x2}
}
\value{
Matrix of similarity calculated from the key and value matrix.
}
\description{
Use some similarity measure to calculate the similarity matrix
}
\details{
Use some similarity measure to calculate the similarity matrix
}
\author{
Luis Gustavo Uzai
}
