% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterEstimation.R
\name{getClusterFact}
\alias{getClusterFact}
\title{Get the Factor of the cluster position in relation to the matrix of eigenvectors}
\usage{
getClusterFact(eigenvectorValues, eigenvectorLengthLessOne, clusterNumber,
  reverseClusterNumber)
}
\arguments{
\item{eigenvectorValues}{Eigenvector matrix based on the affinity matrix}

\item{eigenvectorLengthLessOne}{the eigenvector matrix size minus 1}

\item{clusterNumber}{the cluster position number being tested}

\item{reverseClusterNumber}{the number of the inverse position of the cluster being tested}
}
\value{
Factor of the cluster position in relation to the matrix of eigenvectors
}
\description{
Get the Factor of the cluster position in relation to the matrix of eigenvectors
}
\details{
Gets the factor of the value and its opposite in relation 
          to the matrix of the eigenvectors
}
\author{
Luis Gustavo Uzai
}
