% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectralClustering.R
\name{getSpectralClusters}
\alias{getSpectralClusters}
\title{Clustering with the smallest eigenvectors from eigenvector Matrix}
\usage{
getSpectralClusters(eigenvectorMatrix, numberOfClusters = 2)
}
\arguments{
\item{eigenvectorMatrix}{Eigenvector matrix based on the affinity matrix}

\item{numberOfClusters}{maximum number of clusters for prediction}
}
\value{
K-Means Cluster Object
}
\description{
Clustering with the smallest eigenvectors from eigenvector Matrix
}
\details{
Modified standard function present in kernlab to perform clustering with
         graph spectrum using standard version of K-Means
}
\author{
Luis Gustavo Uzai
}
