% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{\%matches\%.data.frame}
\alias{\%matches\%.data.frame}
\title{Check Whether Data Frame matches Spec}
\usage{
\method{\%matches\%}{data.frame}(x, y, ...)
}
\arguments{
\item{x}{spec}

\item{y}{coerced to spec (spec object or filepath for spec file).}

\item{...}{passed arguments}
}
\value{
logical
}
\description{
Checks whether data.frame matches spec.
}
\examples{
specification(Theoph) \\\%matches\\\% Theoph
}
