% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{\%matches\%.spec}
\alias{\%matches\%.spec}
\title{Check Whether Spec matches Dataset}
\usage{
\method{\%matches\%}{spec}(x, y, ...)
}
\arguments{
\item{x}{spec}

\item{y}{object}

\item{...}{passed arguments}
}
\description{
Checks whether spec matches dataset by checking whether dataset matches spec.
}
\examples{
data(drug)
file <- tempfile()
spec <- specification(drug, tol = 3)
write.spec(spec, file = file)
spec \\\%matches\\\% drug
}
\keyword{internal}
