% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{\%matches\%}
\alias{\%matches\%}
\title{Check Whether x and y Match}
\usage{
x \%matches\% y
}
\arguments{
\item{x}{object}

\item{y}{object}
}
\description{
Checks Whether x an y match.
}
\seealso{
Other matches: 
\code{\link{\%matches\%.character}()},
\code{\link{\%matches\%.data.frame}()},
\code{\link{\%matches\%.spec}()}
}
\concept{matches}
\keyword{internal}
