% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{guidetext}
\alias{guidetext}
\title{Extract Text from Guide}
\usage{
guidetext(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Extracts text from guide.
}
\seealso{
Other guidetext: 
\code{\link{guidetext.spec}()}
}
\concept{guidetext}
\keyword{internal}
