% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{guidetext.spec}
\alias{guidetext.spec}
\title{Extract Text from Guide Column of Specification}
\usage{
\method{guidetext}{spec}(x, column = x$column, ...)
}
\arguments{
\item{x}{object}

\item{column}{character: items in \code{column} to check}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Extracts text from guide column of specification, ignoring number ranges if any.
}
\examples{
data(drug)
guidetext(specification(drug, tol = 3)) # NA, must be supplied manually
}
\seealso{
Other guidetext: \code{\link{guidetext}}
}
\keyword{internal}
