% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar.prob.R
\name{mar.prob}
\alias{mar.prob}
\title{Predictive probability for a marginal test data classification.}
\usage{
mar.prob(x.tr, y.tr, x.te, y.te)
}
\arguments{
\item{x.tr}{Training data vector.}

\item{y.tr}{Training data labels.}

\item{x.te}{Test data vector.}

\item{y.te}{Predicted test data labels}
}
\description{
Predictive probability for a marginal test data classification.
}
\examples{
set.seed(111)
x1<-rPD(10500,10)
x2<-rPD(10500,1000)
test.ind1<-sample.int(10500,500)
test.ind2<-sample.int(10500,500)
x<-c(x1[-test.ind1],x2[-test.ind2])
y1<-rep("1", 10000)
y2<-rep("2", 10000)
y<-c(y1,y2)

t1<-x1[test.ind1]
t2<-x2[test.ind2]
t<-c(t1,t2)

fit<-SPCuPE.fit(x,y)

tM<-tMarLab(fit, t)

mar.prob(x,y,t,tM)
}
