% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_estimation_and_hypothesis_testing.R
\name{MLEp.bsci}
\alias{MLEp.bsci}
\title{Bootstrap confidence interval for the MLE of psi}
\usage{
MLEp.bsci(x, level)
}
\arguments{
\item{x}{A data vector.}

\item{level}{Level of confidence interval as number between 0 and 1.}
}
\value{
The MLE of psi as well as lower and upper bounds of the bootstrap
confidence interval.
}
\description{
A bootstrapped confidence interval for the Maximum Likelihood Estimate for
psi based on a 100 bootstrap rounds with 80\% of data used at a time.
}
\examples{
## Find a 95\% -confidence interval for the MLE of psi given a sample from the
## Poisson-Dirichlet distribution:
x<-rPD(n=10000, psi=100)
MLEp.bsci(x, 0.95)

}
