% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{plot.spectra}
\alias{plot.spectra}
\title{Plot spectra
For now it assumes that all spetra have the the same n bands}
\usage{
\method{plot}{spectra}(this, xlab = "wavelength", ylab = "reflectance",
  xlim = NULL, ylim = NULL, ...)
}
\description{
Plot spectra
For now it assumes that all spetra have the the same n bands
}

