% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{plot.spectrum}
\alias{plot.spectrum}
\title{Plots the reflectance of a spectrum
To "zoom" in part of the spectrum, use xlim = c(min, max)}
\usage{
\method{plot}{spectrum}(this, type = "l", add = FALSE, ...)
}
\arguments{
\item{type}{What type of plot, e.g. "l" or "p"? Passed to "plot.default(type = )".
If add = T, then the only types available are lines and points}

\item{add}{Plot onto an existing plot window? Bool.}

\item{...}{Additional arguments passed to plot.default}
}
\description{
Plots the reflectance of a spectrum
To "zoom" in part of the spectrum, use xlim = c(min, max)
}

