% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{PredictSignature}
\alias{PredictSignature}
\title{Predict the spectral signature of a taxon}
\usage{
PredictSignature(taxon, phy, spectra.df)
}
\arguments{
\item{taxon}{Taxon to estimate (character format)}

\item{phy}{Phylo object with all taxa (including ones with missing data)}

\item{spectra.df}{A data.frame with one row of spectrum observations per species}
}
\value{
A data.frame with the median, upper, and lower estimates for each value for the focal taxon
}
\description{
Predict the spectral signature of a taxon
}

