% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{print.spec}
\alias{print.spec}
\title{Print Spec}
\usage{
\method{print}{spec}(x, limit = 8, ...)
}
\arguments{
\item{x}{spec}

\item{limit}{number of characters to allow without intervention}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Prints spec. Specifically, shortens display of encoded items that are above limit.
}
\concept{print}
\keyword{internal}
