% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{print.spectrum}
\alias{print.spectrum}
\title{Print basic spectrum info}
\usage{
\method{print}{spectrum}(self, full = FALSE)
}
\arguments{
\item{full}{Plot the full object? Bool.}
}
\value{
nothing. Called for its side effect
}
\description{
Print basic spectrum info
}

