% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectrum.R
\name{range.spectrum}
\alias{range.spectrum}
\title{Get the range of spectrum data "wavelength", "reflectance" or "both"
type:}
\usage{
\method{range}{spectrum}(this, type = "wavelength", data = "target")
}
\arguments{
\item{this}{}

\item{type}{Choice of range to return: "wavelength", "reflectance" or "both".}

\item{data}{Choice of data to query: "target" or "reference"}
}
\value{
TODO
}
\description{
Get the range of spectrum data "wavelength", "reflectance" or "both"
type:
}

