% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{respecify.spec}
\alias{respecify.spec}
\title{Respecify Specification}
\usage{
\method{respecify}{spec}(x, data, file = NULL, ...)
}
\arguments{
\item{x}{spec}

\item{data}{a data.frame or path to csv file}

\item{file}{where to write the result (default: do not write)}

\item{...}{passed arguments}
}
\description{
Respecify specification. Updates numeric ranges.  Useful if these have changed and spec no longer matches.
}
