% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPD.R
\name{rPD}
\alias{rPD}
\title{Sampling from he Dirichlet-Poisson Distribution}
\usage{
rPD(n, psi)
}
\arguments{
\item{n}{number of observations.}

\item{psi}{dispersal parameter.}
}
\value{
rPD returns a list with a sample of size n from the Hoppe urn model with parameter psi, along with its table of frequencies.
given parameter psi
}
\description{
rPD samples from the PD distribution by simulating the Hoppe urn model. Time
complexity is something like O(n^2log(psi)). A sample of 10^5 samples
takes 15-20 seconds with a psi<100.
}
\examples{
set.seed(111)
s <- rPD(100,5)
}
\keyword{Poisson-Dirichlet}
\keyword{distribution}
