% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPD.R
\name{rPDa}
\alias{rPDa}
\title{Sampling from he Dirichlet-Poisson Distribution}
\usage{
rPDa(n, psi)
}
\arguments{
\item{n}{number of observations.}

\item{psi}{dispersal parameter.}
}
\value{
rPDa returns a list with a sample of size n from the Hoppe urn model with parameter psi, along with its table of frequencies.
given parameter psi
}
\description{
rPDa samples from the PD distribution by simulating an approximation
of the cylinder set probabilities of PD(psi)
}
\examples{
set.seed(111)
s <- rPDa(100,5)

}
\keyword{Poisson-Dirichlet}
\keyword{distribution}
