% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPCuPE.fit.R
\name{SPCuPE.fit}
\alias{SPCuPE.fit}
\title{Fit the supervised classifier}
\usage{
SPCuPE.fit(x, y)
}
\arguments{
\item{x}{data vector.}

\item{y}{label vector.}
}
\value{
a  list of classwise lists, each with components

frequencies  Frequencies of values

psi          MLE of psi

tr.prob      Probability of the training data under partition exchangeability

coeffs       n!/(psi(psi+1)...(psi+n-1)
}
\description{
The below function trains the model according to training data x and labels y.
The output is a classwise list including the frequencies of the data, the MLE of
psi, the probability of the partition from the logarithm of the Ewens
sampling formula, and a coefficient that equals
$log(n!/(psi(psi+1)...(psi+n-1)).
Saving the frequencies, the training probability and the coefficient save us
from having to calculate these quantities again for the test data.
}
\examples{
set.seed(111)
x1<-rPD(5000,10)
x2<-rPD(5000,100)
x<-c(x1,x2)
y1<-rep("1", 5000)
y2<-rep("2", 5000)
y<-c(y1,y2)
SPCuPE.fit(x,y)
}
\keyword{Fit}
\keyword{data}
\keyword{training}
