% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{specification.comment}
\alias{specification.comment}
\title{Make a Specification for a Comment}
\usage{
\method{specification}{comment}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Makes a specification for a comment
}
\seealso{
Other specification: 
\code{\link{specification.data.frame}()},
\code{\link{specification.default}()},
\code{\link{specification}()}
}
\concept{specification}
\keyword{internal}
