% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{specification.default}
\alias{specification.default}
\title{Make a Specification by Default}
\usage{
\method{specification}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Makes a specification by default method, i.e. returns input unchanged.
}
\seealso{
Other specification: \code{\link{specification.comment}},
  \code{\link{specification.data.frame}},
  \code{\link{specification}}
}
\keyword{internal}
