% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{specification}
\alias{specification}
\title{Make a Specification}
\usage{
specification(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Makes a specification.
}
\examples{
data(drug)
specification(drug, tol = 3)
}
\seealso{
Other specification: \code{\link{specification.comment}},
  \code{\link{specification.data.frame}},
  \code{\link{specification.default}}
}
\keyword{internal}
