% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{specify.character}
\alias{specify.character}
\title{Specify Character}
\usage{
\method{specify}{character}(x, file = sub("csv$", "spec", x),
  spec = read.spec(file), ...)
}
\arguments{
\item{x}{character filepath for a csv file}

\item{file}{character filepath for a matching spec file (ignored if spec provided)}

\item{spec}{a data specification (spec)}

\item{...}{passed arguments}
}
\description{
Attach specifics to a data.frame, supplied as csv filepath.
}
\seealso{
Other specify: \code{\link{specify.data.frame}},
  \code{\link{specify}}
}
