% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{specify}
\alias{specify}
\title{Specify Something}
\usage{
specify(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Specify something.
}
\seealso{
Other specify: \code{\link{specify.character}},
  \code{\link{specify.data.frame}}
}
\keyword{internal}
