% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{SpectraToPhylocurveData}
\alias{SpectraToPhylocurveData}
\title{Convert a data.frame of spectral values (one row per species) to a format suitable for phylocurve}
\usage{
SpectraToPhylocurveData(spectra.df)
}
\arguments{
\item{spectra.df}{A data.frame with one row of spectrum observations per species}
}
\value{
A data.frame with a column for species, wavelength, and measurement at that wavelength
}
\description{
Convert a data.frame of spectral values (one row per species) to a format suitable for phylocurve
}

