% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{[.spectra}
\alias{[.spectra}
\title{Subset operator overload for spectra
TODO! HACK! Learn about this thing ASAP}
\usage{
\method{[}{spectra}(this, i)
}
\description{
Subset operator overload for spectra
TODO! HACK! Learn about this thing ASAP
}

