% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier.R
\name{tMarLab}
\alias{tMarLab}
\title{Marginally predicted labels of the test data given training data classification.}
\usage{
tMarLab(training, x)
}
\arguments{
\item{training}{A training data object from the function SPEC.fit.}

\item{x}{Test data vector or matrix with rows as data points and columns as features.}
}
\value{
A vector of predicted labels for test data x.
}
\description{
tMarLab classifies the test data x based on the training data object.
The test data is considered i.i.d. and to have arrived sequentially. Thus, each
data point is classified one by one.
}
\examples{
set.seed(111)
x1<-rPD(10500,10)
x2<-rPD(10500,1000)
test.ind1<-sample.int(10500,500)
test.ind2<-sample.int(10500,500)
x<-c(x1[-test.ind1],x2[-test.ind2])
y1<-rep("1", 10000)
y2<-rep("2", 10000)
y<-c(y1,y2)

t1<-x1[test.ind1]
t2<-x2[test.ind2]
t<-c(t1,t2)

fit<-SPEC.fit(x,y)

tM<-tMarLab(fit, t)

##With multidimensional x:
set.seed(111)
x1<-cbind(rPD(5500,10),rPD(5500,50))
x2<-cbind(rPD(5500,100),rPD(5500,500))
test.ind1<-sample.int(5500,500)
test.ind2<-sample.int(5500,500)
x<-rbind(x1[-test.ind1,],x2[-test.ind2,])
y1<-rep("1", 5000)
y2<-rep("2", 5000)
y<-c(y1,y2)
fit<-SPEC.fit(x,y)
t1<-x1[test.ind1,]
t2<-x2[test.ind2,]
t<-rbind(t1,t2)

tM<-tMarLab(fit, t)
}
\references{
#' The classification algorithm is adapted from
\link{Corander, J., Cui, Y., Koski, T., and Siren, J.: Have I seen you before? Principles of Bayesian predictive classification revisited. Springer, Stat. Comput. 23, (2011), 59–73.}
(https://doi.org/10.1007/s11222-011-9291-7)
}
\keyword{Marginal}
\keyword{classifier}
