% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neutrality tests.R
\name{WattersonsW}
\alias{WattersonsW}
\title{Watterson's homozygosity test}
\usage{
WattersonsW(x, rounds)
}
\arguments{
\item{x}{A data vector.}

\item{rounds}{How many samples are simulated to obtai nthe empirical distribution.}
}
\value{
A p-value.
}
\description{
This function performs a statistical test on the null hypothesis thata given
sample's underlying distribution is the Poisson-Dirichlet distribution. It
calculates a test statistic that is then used to gain a p-value from an
empirical distribution of the statistic from simulated samples from a
PD distribution.
}
\examples{
##Test whether a typical sample follows PD:
x<-rPD(1000,10)
WattersonsW(x, 1000)

##Test whether a very atypical sample where frequencies of different values
## are similar:

x<-c(rep(1, 200), rep(2, 200), rep(3, 200), rep(4, 200), rep(5,200))
WattersonsW(x,1000)
}
\references{
Watterson, G.A., (1978), The homozygosity test of neutrality. Genetics. 88(2):405-417.
}
