% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{write.spec}
\alias{write.spec}
\title{Write Specification to Storage}
\usage{
write.spec(x, file, ...)
}
\arguments{
\item{x}{spec}

\item{file}{character filepath for storage location}

\item{...}{passed arguments}
}
\description{
Writes specification to storage in tab-delimited format.  Use as.csv() for CSV format.
}
\examples{
data(drug)
file <- tempfile()
spec <- specification(drug, tol = 3)
write.spec(spec, file = file)
}
\seealso{
Other as.spec: \code{\link{as.spec.character}},
  \code{\link{as.spec.data.frame}}, \code{\link{as.spec}},
  \code{\link{read.spec}}
}
