% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abundances}
\alias{abundances}
\title{Vector of frequencies of frequencies}
\usage{
abundances(freqs, freqs0 = NULL)
}
\arguments{
\item{freqs}{A frequency table of data vector \code{x}.}

\item{freqs0}{A second optional frequency table to be merged to the \code{freqs} parameter.}
}
\value{
This function returns a named vector that is used to calculate
probabilities and make classifications.
}
\description{
A function to calculate frequencies of frequencies or the combined
frequencies of frequencies of two vectors.
}
\details{
\code{Freqs0} is optional. It is
used in the case of calculating abundances of test data when the frequencies of
the training data are already known. \code{Freqs0} is \code{table(x0)} ,where \code{x0} is training
data. Abundances of any kind of data vector x can be calculated with
\code{table(table(x))}.
}
\examples{
set.seed(111)
x<-rpois(10,10)
abundances(table(x))

y<-rpois(2,10)
abundances(table(x), table(y))
}
\keyword{abundances}
