% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{as.spec.character}
\alias{as.spec.character}
\title{Coerce to Specification from Character}
\usage{
\method{as.spec}{character}(x, ...)
}
\arguments{
\item{x}{character path to spec-formatted file}

\item{...}{passed arguments}
}
\value{
spec
}
\description{
Coerces to specification from character (length-one filepath).
}
