% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{as.spec.data.frame}
\alias{as.spec.data.frame}
\title{Coerce to Spec from Data Frame}
\usage{
\method{as.spec}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed arguments}
}
\value{
spec
}
\description{
Coerces to spec from data.frame already having basic properties.
}
\examples{
data(drug)
as.spec(specification(drug, tol = 3))
}
\seealso{
Other as.spec: 
\code{\link{as.spec.character}()},
\code{\link{as.spec}()},
\code{\link{read.spec}()},
\code{\link{write.spec}()}
}
\concept{as.spec}
