% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPD.R
\name{dPD}
\alias{dPD}
\title{Poisson Dirichlet distribution}
\usage{
dPD(abund, psi)
}
\arguments{
\item{abund}{An abundance vector.}

\item{psi}{Dispersal parameter. Accepted values are positive numbers, "a" for absolute value psi=1 by default, or "r" for relative value psi equals sample size.}
}
\value{
dPD returns the probability of the abundance vector of the data vector x,
}
\description{
Probability of a data vector x from Ewens' sampling formula. The higher the
dispersal parameter the psi is, the higher the amount of distinct observed
species will be. In terms of the paintbox process, a high psi increases the
size of the continuous part p_0 of the process, while a low psi will increase
the size of the discrete parts p_>0.
}
\examples{
set.seed(111)
s <- rPD(100,5)
a=table(table(s))
dPD(a, 5)

}
\keyword{Poisson-Dirichlet}
\keyword{distribution}
