% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{\%matches\%}
\alias{\%matches\%}
\title{Check Whether x and y Match}
\usage{
x \%matches\% y
}
\arguments{
\item{x}{object}

\item{y}{object}

\item{...}{passed arguments}
}
\description{
Checks Whether x an y match.
}
\keyword{internal}
