% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_estimation_and_hypothesis_testing.R
\name{LMTp}
\alias{LMTp}
\title{Lagrange Multiplier Test for psi}
\usage{
LMTp(abund, psi = "a")
}
\arguments{
\item{abund}{An abundance vector.}

\item{psi}{Target psi to be tested. psi = "a" for absolute value 1, "r" for relative value n (sample size), or any positive number.}
}
\value{
A p-value.
}
\description{
Performs the Lagrange Multiplier test for an abundance vector under partition
exchangeability. Returns a p-value for the hypothesis that the input data
vector stems from a population with the input dispersal parameter.
}
\details{
U(psi_0)^2/I(psi_0),
where U is the log-likelihood function of psi and I is its Fisher information.
The statistic follows chi-squared distribution with 1 degree of freedom
when the null hypothesis H_0:psi<-psi_0 is true.
}
\examples{
set.seed(10000)
x<-rPD(1000, 10)
abund=abundances(table(x))
LMTp(abund, 10)
LMTp(abund, 15)
LMTp(abund, 5)
LMTp(abund)       #test for psi=1
LMTp(abund, "r")  #test for psi=n
}
\keyword{score}
\keyword{test}
