% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEp.bsci.R
\name{MLEp.bsci}
\alias{MLEp.bsci}
\title{Bootstrap confidence interval for the MLE of psi}
\usage{
MLEp.bsci(x)
}
\arguments{
\item{x}{A data vector.}
}
\value{
Lower and upper bounds of the confidence interval.
}
\description{
A bootstrapped confidence interval for the Maximum Likelihood Estimate for psi.
}
