% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEp.R
\name{MLEp}
\alias{MLEp}
\title{Maximum Likelihood Estimate for psi}
\usage{
MLEp(abund)
}
\arguments{
\item{abund}{An abundance vector.}
}
\value{
psi   Maximum Likelihood Estimate of psi

Asymptotic confidence interval
}
\description{
Numerically searches for the MLE of psi as the root of equation
K<-sum(psi/psi+i-1) for i<-1:n, where K is the observed number of
different species in the sample. An accepted psi sets value of the right side
of the equation within R's smallest possible value of the actual value of K.
}
\details{
Numerically searches for the MLE of psi as the root of equation
K<-sum(psi/psi+i-1) for i<-1:n, where K is the observed number of
different species in the sample. An accepted psi sets value of the right side
of the equation within R's smallest possible value of the actual value of K.
}
\examples{
MLEp(abundances(table(c(1,2,2))))

set.seed(1000)
x<-rPD(10000,2000)
MLEp(abundances(table(x)))
}
\keyword{estimate}
\keyword{likelihood}
\keyword{maximum}
\keyword{psi}
