% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_distribution.R
\name{rPD}
\alias{rPD}
\title{Sampling from he Dirichlet-Poisson Distribution}
\usage{
rPD(n, psi)
}
\arguments{
\item{n}{number of observations.}

\item{psi}{dispersal parameter.}
}
\value{
rPD returns a list with a sample of size n from the Hoppe urn model with parameter psi, along with its table of frequencies.
given parameter psi
}
\description{
rPD samples from the PD distribution by simulating the Hoppe urn model.
}
\examples{
set.seed(111)
s <- rPD(100,5)
}
\references{
Hoppe, F.M. The sampling theory of neutral alleles and an urn model in population genetics.
J. Math. Biology 25, 123–159 (1987). https://doi.org/10.1007/BF00276386
}
\keyword{Poisson-Dirichlet}
\keyword{distribution}
