% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_distribution.R
\name{rPD}
\alias{rPD}
\title{Sampling from he Dirichlet-Poisson Distribution}
\usage{
rPD(n, psi)
}
\arguments{
\item{n}{number of observations.}

\item{psi}{dispersal parameter.}
}
\value{
\code{rPD} returns a vector with a sample of size \eqn{n} from the Hoppe urn model with parameter \eqn{\psi}, along with its table of frequencies.
given parameter \eqn{\psi}.
}
\description{
Sampling from he Dirichlet-Poisson Distribution
}
\details{
\code{rPD} samples random values with a given \eqn{\psi} from the Poisson-Dirichlet distribution by simulating the Hoppe urn model.
}
\examples{
## Get random sample from the PD distribution with different psi,
## and estimate the psi of the samples:
s1<-rPD(1000, 10)
s2<- rPD(1000, 50)
print(c(MLEp(table(table(s1)))$psi, MLEp(table(table(s2)))$psi))

}
\references{
Hoppe, F.M. The sampling theory of neutral alleles and an urn model in population genetics.
J. Math. Biology 25, 123–159 (1987). \url{https://doi.org/10.1007/BF00276386}.

W.J. Ewens, The sampling theory of selectively neutral alleles, Theoretical Population Biology, Volume 3, Issue 1,
1972, Pages 87-112, ISSN 0040-5809, \url{https://doi.org/10.1016/0040-5809(72)90035-4}.
}
\keyword{Poisson-Dirichlet}
\keyword{distribution}
