% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier.R
\name{SPEC.fit}
\alias{SPEC.fit}
\title{Fit the supervised classifier}
\usage{
SPEC.fit(x, y)
}
\arguments{
\item{x}{data vector, or matrix with rows as data points and columns as features.}

\item{y}{training label vector.}
}
\value{
If x is multidimensional, each list described below is returned for each dimension.

Returns a list of classwise lists, each with components:

frequencies: the frequencies of values in the class.

psi: the estimate of psi for the class.
}
\description{
Trains the model according to training data x and labels y.
The output is a classwise list including the frequencies of the data, and the MLE of
psi.
}
\examples{
set.seed(111)
x1<-rPD(5000,10)
x2<-rPD(5000,100)
x<-c(x1,x2)
y1<-rep("1", 5000)
y2<-rep("2", 5000)
y<-c(y1,y2)
fit<-SPEC.fit(x,y)

##With multidimensional x:
set.seed(111)
x1<-cbind(rPD(5000,10),rPD(5000,50))
x2<-cbind(rPD(5000,100),rPD(5000,500))
x<-rbind(x1,x2)
y1<-rep("1", 5000)
y2<-rep("2", 5000)
y<-c(y1,y2)
fit<-SPEC.fit(x,y)
}
\keyword{Fit}
\keyword{data}
\keyword{training}
