% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tMargLab.R
\name{tMarLab}
\alias{tMarLab}
\title{Marginally predicted labels of the test data given training data classification.}
\usage{
tMarLab(training, x)
}
\arguments{
\item{training}{A training data object from the function SPCuPE.fit.}

\item{x}{Test data vector.}
}
\value{
A vector of predicted labels for test data x.
}
\description{
tMarLab classifies the test data x based on the training data object.
The test data is considered i.i.d. and to have arrived sequentially. Thus, each
data point is classified one by one.
The classification algorithm is from
\link{Corander, J., Cui, Y., Koski, T., and Siren, J.: Have I seen you before? Principles of Bayesian predictive classification revisited. Springer, Stat. Comput. 23, (2011), 59–73.}
(https://doi.org/10.1007/s11222-011-9291-7)
}
\examples{
set.seed(111)
x1<-rPD(10500,10)
x2<-rPD(10500,1000)
test.ind1<-sample.int(10500,500)
test.ind2<-sample.int(10500,500)
x<-c(x1[-test.ind1],x2[-test.ind2])
y1<-rep("1", 10000)
y2<-rep("2", 10000)
y<-c(y1,y2)

t1<-x1[test.ind1]
t2<-x2[test.ind2]
t<-c(t1,t2)

fit<-SPCuPE.fit(x,y)

tM<-tMarLab(fit, t)


v<-c(x1[1:10000],x2[1:10000])
z<-c(x1[10001:10500],x2[10001:10500])

fit2<-SPCuPE.fit(v,y)
tM2<-tMarLab(fit2,z)
}
\keyword{Marginal}
\keyword{classifier}
